# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: WASAText messaging service API for Web and Software Architecture course
  version: "1.0.0"
...
servers:
  - url: http://localhost:3000
paths:
...
  /users/{userId}/username:
    parameters:
      - $ref: "#/components/parameters/userId"
    put:
      tags: ["User"]
      operationId: setMyUserName
      summary: Change username
      description: Sets a new username for the user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for setting username
              required:
                - username
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 16
                  pattern: "^[a-zA-Z0-9_]+$"
      responses:
        "204":
          description: Username changed successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Username already taken
        "500":
          $ref: "#/components/responses/InternalServerError"
  /users/{userId}/photo:
    parameters:
      - $ref: "#/components/parameters/userId"
    put:
      tags: ["User"]
      operationId: setMyPhoto
      summary: Set user profile photo
      description: Uploads a new profile photo for the user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: Binary PNG image data
          image/jpeg:
            schema:
              type: string
              format: binary
              description: Binary JPEG image data
      responses:
        "204":
          description: Photo updated successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/conversationId"
    post:
      tags: ["Messages"]
      operationId: sendMessage
      summary: Send a message
      description: Sends a text or photo message to the conversation. Can optionally be a reply.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: JSON body for sending a message
              required:
                - type
              properties:
                type:
                  type: string
                  enum: ["text", "photo"]
                  description: Message type
                content:
                  type: string
                  description: Text content (required if type is text)
                  minLength: 1
                  maxLength: 4096
                replyToId:
                  type: string
                  description: Message ID being replied to (optional)
                  minLength: 1
                  maxLength: 64
          multipart/form-data:
            schema:
            schema:
              type: object
              description: Multipart body for sending a photo
              properties:
                type:
                  type: string
                  enum: ["photo"]
                  description: Message type
                photo:
                  type: string
                  format: binary
                  description: Binary image data
                replyToId:
                  type: string
                  description: Message ID being replied to
                  minLength: 1
                  maxLength: 64
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not a member of this conversation
        "500":
          $ref: "#/components/responses/InternalServerError"
  /conversations/{conversationId}/messages/forward:
    parameters:
      - $ref: "#/components/parameters/conversationId"
    post:
      tags: ["Messages"]
      operationId: forwardMessage
      summary: Forward a message
      description: Forwards an existing message to this conversation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for forwarding a message
              required:
                - messageId
              properties:
                messageId:
                  type: string
                  description: ID of the message to forward
                  minLength: 1
                  maxLength: 64
      responses:
        "201":
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not a member of this conversation
        "404":
          description: Original message not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/messageId"
    delete:
      tags: ["Messages"]
      operationId: deleteMessage
      summary: Delete a message
      description: Deletes a message. Only the sender can delete their own messages.
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Message deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: User is not the sender of this message
        "404":
          description: Message not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /messages/{messageId}/comment:
    parameters:
      - $ref: "#/components/parameters/messageId"
    put:
      tags: ["Reactions"]
      operationId: commentMessage
      summary: Add a comment/reaction to a message
      description: Adds or updates the user's comment on a message. Each user can have only one comment per message.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON body for adding a comment
              required:
                - comment
              properties:
                comment:
                  type: string
                  maxLength: 50
                  description: The comment or reaction emoji
      responses:
        "204":
          description: Comment added successfully
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Message not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: ["Reactions"]
      operationId: uncommentMessage
      summary: Remove a comment/reaction from a message
      description: Removes the user's comment from a message
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Comment removed successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Message or comment not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /users:
    get:
      tags: ["User"]
      operationId: searchUsers
      summary: Search users
      description: Search for users by username to start conversations
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query (username substring)
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: List of matching users
          content:
            application/json:
              schema:
                type: array
                minItems: 0
                maxItems: 100
                items:
                  $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: User identifier returned from doLogin
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation identifier
      schema:
        type: string
    messageId:
      name: messageId
      in: path
      required: true
      description: Message identifier
      schema:
        type: string
...
  schemas:
...
    User:
      type: object
      description: Represents a user in the system
      properties:
        id:
          type: string
          description: Unique identifier of the user
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: The user's display name
          minLength: 3
          maxLength: 16
          pattern: "^[a-zA-Z0-9_]+$"
        photoUrl:
          type: string
          description: URL to the user's profile photo
          nullable: true
          maxLength: 2048
      required:
        - id
        - username
...
    MessagePreview:
      type: object
      description: A summary of a message
      properties:
        content:
          type: string
          description: The content of the message
          minLength: 1
          maxLength: 4096
        timestamp:
          type: string
          description: The time the message was sent
          format: date-time
        senderId:
          type: string
          description: ID of the user who sent the message
          minLength: 1
          maxLength: 64
      required:
        - content
        - timestamp
        - senderId
...
    Message:
      type: object
      description: A message within a conversation
      properties:
        id:
          type: string
          description: Unique identifier of the message
          minLength: 1
          maxLength: 64
        senderId:
          type: string
          description: ID of the sender
          minLength: 1
          maxLength: 64
        senderUsername:
          type: string
          description: Username of the sender
          minLength: 3
          maxLength: 16
        type:
          type: string
          description: Type of message content
          enum: ["text", "photo"]
        content:
          type: string
          description: Text content or photo URL
          minLength: 1
          maxLength: 4096
        timestamp:
          type: string
          description: ISO 8601 timestamp
          format: date-time
        checkmarks:
          type: integer
          enum: [0, 1, 2]
          description: "0=sent, 1=received by all, 2=read by all"
        replyTo:
          allOf:
            - $ref: "#/components/schemas/MessagePreview"
            - nullable: true
        forwarded:
          type: boolean
          description: Whether the message was forwarded
        comments:
          type: array
          description: List of reactions/comments
          items:
            $ref: "#/components/schemas/Comment"
          minItems: 0
          maxItems: 100
      required:
        - id
        - senderId
        - senderUsername
        - type
        - content
        - timestamp
        - checkmarks
        - forwarded
        - comments
    Comment:
      type: object
      description: A reaction or comment on a message
      properties:
        userId:
          type: string
          description: ID of the user who commented
          minLength: 1
          maxLength: 64
        username:
          type: string
          description: Username of the commenter
          minLength: 3
          maxLength: 16
        comment:
          type: string
          description: The comment text or emoji
          minLength: 1
          maxLength: 50
      required:
        - userId
        - username
        - comment
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation
    Unauthorized:
      description: The access token is missing or invalid
    InternalServerError:
      description: The server encountered an internal error